requirements.txt
fastapi
uvicorn
groq
dspy-ai
python-dotenv
pydantic
supabase
google-genai
